Make tsan an optional extra
The tsan switch takes longer to make and uses a fair bit more memory.
In order to try to ensure the container builds successfully, comment
out the tsan switch creation. Once we get to the part of the tutorial
that involves tsan we can suggest people try it out at that point.